requirements.txt
certifi==2020.6.20
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn-pandas==1.5.0
threadpoolctl==3.1.0
typing_extensions==4.1.1
Werkzeug==2.0.3
wincertstore==0.2
xgboost==1.5.2
zipp==3.6.0